+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 All new features, changes and fixes should be listed here. Please use tickets to reference changes.
 
-## 0.1.0 (2014/xx/xx)
+## 0.1.0 (2014/06/28)
 
-* [new] initial version
+* [add] ```search($expression)``` method to get multiple values at once or retrieve nested values via [JMESPath expressions](http://jmespath.readthedocs.org/en/latest/index.html) (#2)
+* [add] Parameters array wrapper class with conveninient `get($key, $default)` that implements `\ArrayAccess` (#1)
 * [fix] none
 * [chg] none